refactor: Apply SRP principal to the OffsetManagerRepository
- Delegate storage layer to the OffsetStorage and does synchronization to the OffsetStateUpdater
