+++ offset-manager/src/test/java/offsetmanager/repository/OffsetManagerRepositoryTest.java
@@ -1,11 +1,14 @@
 package offsetmanager.repository;
 
 import offsetmanager.domain.InMemoryOffsetStorage;
+import offsetmanager.domain.OffsetStateUpdater;
+import offsetmanager.domain.OffsetStateUpdaterImpl;
 import offsetmanager.domain.OffsetStorage;
 import offsetmanager.domain.file.FileKey;
 import offsetmanager.domain.file.factory.FileKeyParser;
 import offsetmanager.domain.offset.DefaultOffsetRecord;
 import offsetmanager.domain.offset.OffsetRecord;
+import offsetmanager.exception.OffsetManagerNotReadyException;
 import offsetmanager.support.KafkaTestSupport;
 import org.apache.kafka.clients.producer.Producer;
 import org.apache.kafka.clients.producer.ProducerRecord;
@@ -18,6 +21,7 @@ import java.util.List;
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class OffsetManagerRepositoryTest extends KafkaTestSupport {
 
@@ -41,21 +45,33 @@ class OffsetManagerRepositoryTest extends KafkaTestSupport {
   @Test
   void emptyOffsetTopicTest() {
     // given
-    OffsetManagerRepository offsetManager = new OffsetManagerRepository(offsetStorage, testConsumerProperties, this.offsetTopic);
+    OffsetStateUpdater offsetStateUpdater = new OffsetStateUpdaterImpl(offsetTopic, testConsumerProperties, offsetStorage);
+    offsetStateUpdater.start();
+    OffsetManagerRepository offsetManager = new OffsetManagerRepository(offsetStorage, offsetStateUpdater);
     // when
     Optional<OffsetRecord> foundOffset = offsetManager.findLatestOffsetRecord(FileKeyParser.parse("file:///test/path.txt"));
     // then
     assertThat(foundOffset).isEmpty();
+  }
 
-    // cleans
-    offsetManager.shutdown();
+  @DisplayName("Should throw OffsetManagerNotReadyException when OffsetStateUpdater not started")
+  @Test
+  void offsetUpdaterNotReadyTest() {
+    // given
+    OffsetStateUpdater offsetStateUpdater = new OffsetStateUpdaterImpl(offsetTopic, testConsumerProperties, offsetStorage);
+    OffsetManagerRepository offsetManager = new OffsetManagerRepository(offsetStorage, offsetStateUpdater);
+    // when then
+    assertThatThrownBy(()-> offsetManager.findLatestOffsetRecord(FileKeyParser.parse("file:///test/path.txt")))
+      .isInstanceOf(OffsetManagerNotReadyException.class);
   }
 
   @DisplayName("Should find all offset records by keys")
   @Test
   void findAllOffsetRecordsTest() throws InterruptedException {
     // given
-    OffsetManagerRepository offsetManager = new OffsetManagerRepository(offsetStorage, testConsumerProperties, this.offsetTopic);
+    OffsetStateUpdater offsetStateUpdater = new OffsetStateUpdaterImpl(offsetTopic, testConsumerProperties, offsetStorage);
+    offsetStateUpdater.start();
+    OffsetManagerRepository offsetManager = new OffsetManagerRepository(offsetStorage, offsetStateUpdater);
     FileKey keyA = FileKeyParser.parse("file:///many-a.txt");
     FileKey keyB = FileKeyParser.parse("file:///many-b.txt");
     FileKey keyC = FileKeyParser.parse("file:///many-c.txt");
@@ -84,14 +100,16 @@ class OffsetManagerRepositoryTest extends KafkaTestSupport {
       );
 
     // cleans
-    offsetManager.shutdown();
+    offsetStateUpdater.stop();
   }
 
   @DisplayName("Update continuously receives new offsets and updates the store")
   @Test
   void upsertContinuously() throws InterruptedException {
     // given
-    OffsetManagerRepository offsetManager = new OffsetManagerRepository(offsetStorage, testConsumerProperties, this.offsetTopic);
+    OffsetStateUpdater offsetStateUpdater = new OffsetStateUpdaterImpl(offsetTopic, testConsumerProperties, offsetStorage);
+    offsetStateUpdater.start();
+    OffsetManagerRepository offsetManager = new OffsetManagerRepository(offsetStorage, offsetStateUpdater);
     FileKey keyA = FileKeyParser.parse("file:///key-a.txt");
     FileKey keyB = FileKeyParser.parse("file:///key-b.txt");
     FileKey keyC = FileKeyParser.parse("file:///key-c.txt");
@@ -120,6 +138,6 @@ class OffsetManagerRepositoryTest extends KafkaTestSupport {
       .isEqualTo(new DefaultOffsetRecord(keyC, 1000L));
 
     // cleans
-    offsetManager.shutdown();
+    offsetStateUpdater.stop();
   }
 }
